Refactor and simplify
