+++ packages/studio/src/application/modules/animation/ui/AnimationTimelines.tsx
@@ -107,7 +107,7 @@ export const AnimationTimelines: FC<AnimationTimelinesProps> = ({
           </MenuRadioGroup>
         </Menu>
         <ToolSpacer />
-        <ToolButton disabled icon={<Icon>?</Icon>} tooltip="Help" />
+        <ToolButton disabled icon={<Icon colorize>?</Icon>} tooltip="Help" />
       </ToolBar>
       <AnimationsContainer
         duration={(maxTime + EXTRA_TIME) / 1000}
